# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vital Track API - calorie, steps, and exercise tracker
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Local account and session operations
  - name: calories
    description: Calorie log operations
  - name: steps
    description: Step log operations
  - name: exercises
    description: Exercise session operations
  - name: summary
    description: Summary and analytics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: registerAccount
      tags: [auth]
      summary: Create a local account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the current session
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Signed out

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the signed-in user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: "#/components/schemas/AuthUser"

  /calories:
    get:
      operationId: listCalorieLogs
      tags: [calories]
      summary: List calorie log entries
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of calorie entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalorieLog"
    post:
      operationId: createCalorieLog
      tags: [calories]
      summary: Log a food/calorie entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalorieLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalorieLog"

  /calories/{id}:
    get:
      operationId: getCalorieLog
      tags: [calories]
      summary: Get a calorie log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Calorie log entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalorieLog"
        "404":
          description: Not found
    patch:
      operationId: updateCalorieLog
      tags: [calories]
      summary: Update a calorie log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalorieLogUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalorieLog"
    delete:
      operationId: deleteCalorieLog
      tags: [calories]
      summary: Delete a calorie log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /steps:
    get:
      operationId: listStepLogs
      tags: [steps]
      summary: List step log entries
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of step entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StepLog"
    post:
      operationId: createStepLog
      tags: [steps]
      summary: Log steps for a day
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepLog"

  /steps/{id}:
    patch:
      operationId: updateStepLog
      tags: [steps]
      summary: Update a step log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepLogUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepLog"
    delete:
      operationId: deleteStepLog
      tags: [steps]
      summary: Delete a step log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List exercise sessions
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of exercise sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Log an exercise session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"

  /exercises/{id}:
    patch:
      operationId: updateExercise
      tags: [exercises]
      summary: Update an exercise session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
    delete:
      operationId: deleteExercise
      tags: [exercises]
      summary: Delete an exercise session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /summary/today:
    get:
      operationId: getTodaySummary
      tags: [summary]
      summary: Get today's summary - calories, steps, exercises
      responses:
        "200":
          description: Today's summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummary"

  /summary/weekly:
    get:
      operationId: getWeeklySummary
      tags: [summary]
      summary: Get the last 7 days summary for charts
      responses:
        "200":
          description: Weekly summary data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DayStats"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthUser:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email

    AuthSession:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"

    CalorieLog:
      type: object
      required: [id, foodName, calories, mealType, loggedAt]
      properties:
        id:
          type: integer
        foodName:
          type: string
        calories:
          type: integer
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        loggedAt:
          type: string

    CalorieLogInput:
      type: object
      required: [foodName, calories, mealType]
      properties:
        foodName:
          type: string
          minLength: 1
        calories:
          type: integer
          minimum: 0
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        loggedAt:
          type: string

    CalorieLogUpdate:
      type: object
      properties:
        foodName:
          type: string
        calories:
          type: integer
        protein:
          type: ["number", "null"]
        carbs:
          type: ["number", "null"]
        fat:
          type: ["number", "null"]
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]

    StepLog:
      type: object
      required: [id, steps, date, loggedAt]
      properties:
        id:
          type: integer
        steps:
          type: integer
        date:
          type: string
        distanceKm:
          type: ["number", "null"]
        caloriesBurned:
          type: ["integer", "null"]
        loggedAt:
          type: string

    StepLogInput:
      type: object
      required: [steps, date]
      properties:
        steps:
          type: integer
          minimum: 0
        date:
          type: string
        distanceKm:
          type: ["number", "null"]
        caloriesBurned:
          type: ["integer", "null"]

    StepLogUpdate:
      type: object
      properties:
        steps:
          type: integer
        distanceKm:
          type: ["number", "null"]
        caloriesBurned:
          type: ["integer", "null"]

    Exercise:
      type: object
      required: [id, name, category, durationMinutes, loggedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [cardio, strength, flexibility, sports, other]
        durationMinutes:
          type: integer
        caloriesBurned:
          type: ["integer", "null"]
        sets:
          type: ["integer", "null"]
        reps:
          type: ["integer", "null"]
        weightKg:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        loggedAt:
          type: string

    ExerciseInput:
      type: object
      required: [name, category, durationMinutes]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum: [cardio, strength, flexibility, sports, other]
        durationMinutes:
          type: integer
          minimum: 1
        caloriesBurned:
          type: ["integer", "null"]
        sets:
          type: ["integer", "null"]
        reps:
          type: ["integer", "null"]
        weightKg:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        loggedAt:
          type: string

    ExerciseUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [cardio, strength, flexibility, sports, other]
        durationMinutes:
          type: integer
        caloriesBurned:
          type: ["integer", "null"]
        sets:
          type: ["integer", "null"]
        reps:
          type: ["integer", "null"]
        weightKg:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]

    DailySummary:
      type: object
      required: [date, totalCaloriesIn, totalCaloriesBurned, totalSteps, totalDistanceKm, totalExerciseMinutes, calorieGoal, stepGoal, netCalories]
      properties:
        date:
          type: string
        totalCaloriesIn:
          type: integer
        totalCaloriesBurned:
          type: integer
        totalSteps:
          type: integer
        totalDistanceKm:
          type: number
        totalExerciseMinutes:
          type: integer
        calorieGoal:
          type: integer
        stepGoal:
          type: integer
        netCalories:
          type: integer
        mealBreakdown:
          type: object
          properties:
            breakfast:
              type: integer
            lunch:
              type: integer
            dinner:
              type: integer
            snack:
              type: integer

    DayStats:
      type: object
      required: [date, caloriesIn, caloriesBurned, steps, distanceKm, exerciseMinutes]
      properties:
        date:
          type: string
        caloriesIn:
          type: integer
        caloriesBurned:
          type: integer
        steps:
          type: integer
        distanceKm:
          type: number
        exerciseMinutes:
          type: integer