# lib/api-zod/src/generated/types/dailySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vital Track API - calorie, steps, and exercise tracker
 * OpenAPI spec version: 0.1.0
 */
import type { DailySummaryMealBreakdown } from './dailySummaryMealBreakdown';

export interface DailySummary {
  date: string;
  totalCaloriesIn: number;
  totalCaloriesBurned: number;
  totalSteps: number;
  totalDistanceKm: number;
  totalExerciseMinutes: number;
  calorieGoal: number;
  stepGoal: number;
  netCalories: number;
  mealBreakdown?: DailySummaryMealBreakdown;
}