# lib/api-zod/src/generated/types/dayStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vital Track API - calorie, steps, and exercise tracker
 * OpenAPI spec version: 0.1.0
 */

export interface DayStats {
  date: string;
  caloriesIn: number;
  caloriesBurned: number;
  steps: number;
  distanceKm: number;
  exerciseMinutes: number;
}