# lib/api-zod/src/generated/types/exerciseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vital Track API - calorie, steps, and exercise tracker
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseInputCategory } from './exerciseInputCategory';

export interface ExerciseInput {
  /** @minLength 1 */
  name: string;
  category: ExerciseInputCategory;
  /** @minimum 1 */
  durationMinutes: number;
  /** @nullable */
  caloriesBurned?: number | null;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  notes?: string | null;
  loggedAt?: string;
}