# lib/api-zod/src/generated/types/exerciseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vital Track API - calorie, steps, and exercise tracker
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseUpdateCategory } from './exerciseUpdateCategory';

export interface ExerciseUpdate {
  name?: string;
  category?: ExerciseUpdateCategory;
  durationMinutes?: number;
  /** @nullable */
  caloriesBurned?: number | null;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  notes?: string | null;
}